feature pizza main list module
