+++ app/src/main/java/com/example/pizzashift2025/di/ViewModelModule.kt
@@ -1,7 +1,12 @@
 package com.example.pizzashift2025.di
 
-import com.example.pizzashift2025.pizza_main.di.ViewModelPizzaMainModule
+import androidx.lifecycle.ViewModel
+import com.example.pizzashift2025.presentation.MainViewModel
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.di.ViewModelPizzaMainModule
+import com.example.pizzashift2025.util.ViewModelKey
+import dagger.Binds
 import dagger.Module
+import dagger.multibindings.IntoMap
 
 @Module(
     includes = [
@@ -9,4 +14,8 @@ import dagger.Module
     ]
 )
 interface ViewModelModule {
+    @Binds
+    @IntoMap
+    @ViewModelKey(MainViewModel::class)
+    fun bindViewModelMain(viewModel: MainViewModel): ViewModel
 }
