+++ app/src/main/java/com/example/pizzashift2025/navigation/di/NavigationModule.kt
@@ -4,8 +4,8 @@ import com.example.pizzashift2025.navigation.GlobalRouter
 import com.example.pizzashift2025.navigation.GlobalRouterImpl
 import com.example.pizzashift2025.navigation.featureRoute.MainListRouterImpl
 import com.example.pizzashift2025.navigation.featureRoute.MainRouterImpl
-import com.example.pizzashift2025.pizza_main.presentation.MainListRouter
-import com.example.pizzashift2025.pizza_main.presentation.MainRouter
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.presentation.MainListRouter
+import com.example.pizzashift2025.presentation.MainRouter
 import dagger.Module
 import dagger.Provides
