+++ app/src/main/java/com/example/pizzashift2025/navigation/featureRoute/MainListRouterImpl.kt
@@ -1,7 +1,7 @@
 package com.example.pizzashift2025.navigation.featureRoute
 
 import com.example.pizzashift2025.navigation.GlobalRouter
-import com.example.pizzashift2025.pizza_main.presentation.MainListRouter
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.presentation.MainListRouter
 import javax.inject.Inject
 
 class MainListRouterImpl @Inject constructor(
