+++ app/src/main/java/com/example/pizzashift2025/navigation/featureRoute/MainRouterImpl.kt
@@ -1,8 +1,8 @@
 package com.example.pizzashift2025.navigation.featureRoute
 
 import com.example.pizzashift2025.navigation.GlobalRouter
-import com.example.pizzashift2025.pizza_main.presentation.MainListRoute
-import com.example.pizzashift2025.pizza_main.presentation.MainRouter
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.MainListRoute
+import com.example.pizzashift2025.presentation.MainRouter
 import javax.inject.Inject
 
 class MainRouterImpl @Inject constructor(
