--- app/src/main/java/com/example/pizzashift2025/pizza_main/presentation/MainRouter.kt
@@ -1,5 +1,0 @@
-package com.example.pizzashift2025.pizza_main.presentation
-
-interface MainRouter {
-    fun openListPizza()
-}
