--- app/src/main/java/com/example/pizzashift2025/pizza_main/presentation/NavigationOption.kt
@@ -1,7 +1,0 @@
-package com.example.pizzashift2025.pizza_main.presentation
-
-import kotlin.reflect.KClass
-
-enum class NavigationOption(val route: KClass<*>) {
-    CATALOG(MainListRoute::class)
-}
