+++ app/src/main/java/com/example/pizzashift2025/presentation/MainRouter.kt
@@ -1,0 +1,5 @@
+package com.example.pizzashift2025.presentation
+
+interface MainRouter {
+    fun openListPizza()
+}
