+++ app/src/main/java/com/example/pizzashift2025/presentation/MainState.kt
@@ -1,4 +1,4 @@
-package com.example.pizzashift2025.pizza_main.presentation
+package com.example.pizzashift2025.presentation
 
 data class MainState(
     val navigationOption: List<NavigationOption>,
