+++ app/src/main/java/com/example/pizzashift2025/presentation/NavigationOption.kt
@@ -1,0 +1,8 @@
+package com.example.pizzashift2025.presentation
+
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.MainListRoute
+import kotlin.reflect.KClass
+
+enum class NavigationOption(val route: KClass<*>) {
+    CATALOG(MainListRoute::class)
+}
