+++ app/src/main/java/com/example/pizzashift2025/ui/PizzaMainScreen.kt
@@ -1,4 +1,4 @@
-package com.example.pizzashift2025.pizza_main.ui
+package com.example.pizzashift2025.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Surface
@@ -9,8 +9,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.pizzashift2025.pizza_main.presentation.MainListRoute
-import com.example.pizzashift2025.pizza_main.presentation.MainViewModel
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.MainListRoute
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.ui.PizzaListScreen
+import com.example.pizzashift2025.presentation.MainViewModel
 import com.example.pizzashift2025.util.getViewModel
 
 @Composable
