+++ feature/pizza_main_list/.gitignore
@@ -1,0 +1,2 @@
+/build
+/.idea
