+++ feature/pizza_main_list/build.gradle.kts
@@ -1,36 +1,20 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.jetbrains.kotlin.android)
     alias(libs.plugins.jetbrains.kotlin.serialization)
     alias(libs.plugins.kotlin.kapt)
 }
 
 android {
-    namespace = "com.example.pizzashift2025"
+    namespace = "com.example.pizzashift2025.feature.pizza_main_list"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.example.pizzashift2025"
         minSdk = 24
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
-
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
+
     }
 
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
@@ -52,13 +36,13 @@ android {
 }
 
 dependencies {
-    implementation(project(":util"))
-    implementation(project(":design:resources"))
     implementation(project(":shared:pizza"))
-    implementation(project(":feature:pizza_main_list"))
+    implementation(project(":util"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
+    implementation(libs.androidx.lifecycle.viewmodel.compose.android)
+    implementation(libs.androidx.lifecycle.runtime.compose.android)
 
     // Compose
     implementation(libs.androidx.activity.compose)
@@ -73,9 +57,6 @@ dependencies {
     // Serialization
     implementation(libs.kotlinx.serialization.json)
 
-    // Navigation
-    implementation(libs.androidx.navigation.compose)
-
     // Kotlin Coroutines
     implementation(libs.kotlinx.coroutines.core)
     implementation (libs.kotlinx.coroutines.android)
@@ -83,14 +64,6 @@ dependencies {
     // Dagger
     implementation(libs.dagger)
     kapt(libs.dagger.compiler)
-
-    //Retrofit
-    implementation(libs.retrofit)
-    implementation(libs.converter.gson)
-
-    // OkHttp
-    implementation (libs.okhttp)
-    implementation(libs.logging.interceptor)
 
     // Coil Compose
     implementation (libs.coil.kt.coil.compose)
