+++ feature/pizza_main_list/src/main/kotlin/com/example/pizzashift2025/feature/pizza_main_list/pizza_main_list/MainListRoute.kt
@@ -1,4 +1,4 @@
-package com.example.pizzashift2025.pizza_main.presentation
+package com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list
 
 import kotlinx.serialization.Serializable
