+++ feature/pizza_main_list/src/main/kotlin/com/example/pizzashift2025/feature/pizza_main_list/pizza_main_list/di/ViewModelPizzaMainModule.kt
@@ -1,9 +1,8 @@
-package com.example.pizzashift2025.pizza_main.di
+package com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.di
 
 import androidx.lifecycle.ViewModel
 import com.example.pizzashift2025.util.ViewModelKey
-import com.example.pizzashift2025.pizza_main.presentation.MainViewModel
-import com.example.pizzashift2025.pizza_main.presentation.PizzaListMainViewModel
+import com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.presentation.PizzaListMainViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
@@ -14,9 +13,4 @@ interface ViewModelPizzaMainModule {
     @IntoMap
     @ViewModelKey(PizzaListMainViewModel::class)
     fun bindViewModelPizzaListMain(viewModel: PizzaListMainViewModel): ViewModel
-
-    @Binds
-    @IntoMap
-    @ViewModelKey(MainViewModel::class)
-    fun bindViewModelMain(viewModel: MainViewModel): ViewModel
 }
