+++ feature/pizza_main_list/src/main/kotlin/com/example/pizzashift2025/feature/pizza_main_list/pizza_main_list/presentation/MainListRouter.kt
@@ -1,0 +1,5 @@
+package com.example.pizzashift2025.feature.pizza_main_list.pizza_main_list.presentation
+
+interface MainListRouter {
+    fun openPizza(id: Int?)
+}
