+++ feature/pizza_main_list/src/main/res/values-ru/strings.xml
