+++ feature/pizza_main_list/src/main/res/values/strings.xml
